Added testing functions. Fixed data incompatibility errors.

+++ src/Utils/SMS.java
@@ -7,7 +7,7 @@ public class SMS implements INotification{
     public boolean sendNotification(Transaction transaction) {
         String currAccName=transaction.getCurrAccount().getUser().getName();
         String otherAccName=transaction.getOtherAccount().getUser().getName();
-        float amt = transaction.getAmount();
+        double amt = transaction.getAmount();
         //implement external API call to send SMS
         if(transaction.getType()== TransactionType.CREDIT)
             System.out.println(currAccName+ " was credited by "+amt+" from "+otherAccName);
